test: Move new config-type of test to test_config.py
