+++ tests/test_config_file.py
@@ -321,21 +321,3 @@ def test_merge_nested_dict(tmpdir, config_dir):
     assert str(config2) in configs
     with pytest.raises(Exception):
         config.load_configs(configs)
-
-
-def test_relative_dir(tmpdir):
-    arbitrary_dir = tmpdir.join("moo")
-    arbitrary_dir.mkdir()
-
-    arbitrary_dir.join("rel.yaml").write(
-        """
-./relativedir:
-  docutils: svn+http://svn.code.sf.net/p/docutils/code/trunk
-   """
-    )
-
-    configs = config.find_config_files(path=str(arbitrary_dir))
-    repos = config.load_configs(configs, str(arbitrary_dir))
-
-    assert str(arbitrary_dir.join("relativedir")) == repos[0]["parent_dir"]
-    assert str(arbitrary_dir.join("relativedir", "docutils")) == repos[0]["repo_dir"]
